updated type locations
